Changes
Instead of prompting the user to provide a URL for the image, it will now prompt the user to attach a local image like when initially adding a tile.

When using the "18+" tag users will now have the ability to blur the images. Right now the export nor import functionality do not have the ability to save or import the state which the button is in.