Update character.html
Now instead of the new tiles appearing before the add-image tile, it should now be added after the latest tile. This should in theory allow the movement of the add image tile without affecting where the placement of the new tiles are. Also changed theme buttons to one singular button